Big clean up and rename
