+++ parsing/byte/generate.go
