+++ parsing/byte_stream.go
@@ -1,4 +1,4 @@
-package parser
+package parsing
 
 import (
 	"bufio"
