+++ parsing/byte_stream_test.go
@@ -1,4 +1,4 @@
-package parser
+package parsing
 
 import (
 	"bytes"
