+++ parsing/syntax_error.go
@@ -1,4 +1,4 @@
-package parser
+package parsing
 
 import (
 	"fmt"
